rmmod: prevent error if module directory does not exists.
Modules with rmmod are removed by name and it is not important if a module is in
the default module's path or if the path exist, so it useless check if default
path exist and require its existence.

No error is displayed when you try to unload a nonexistent module; only if
debug mode is active a message is showed. It can be a problem because on
missed typeing error busybox do not return error and it might be thought that
the real module is removed, but it not and it might be a problem.

Signed-off-by: Federico Vaga <federico.vaga@gmail.com>